requirements.txt
python-dotenv==0.20.0
requests==2.28.1
urllib3==1.26.11
